Consolidate localize error functions
